+++ core/src/main/scala/cats/tagless/InvariantK.scala
@@ -16,7 +16,7 @@
 
 package cats.tagless
 
-import cats.hkinds.instances.InvariantKInstances
+import cats.tagless.instances.InvariantKInstances
 import cats.~>
 import simulacrum.typeclass
