+++ core/src/main/scala/cats/tagless/instances/ContravariantKInstances.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package cats.hkinds.instances
+package cats.tagless.instances
 
 import cats._
 import cats.kernel.CommutativeMonoid
